fix: use makeInformer
